# contracts/openapi/troubleshooting-api.yaml
openapi: 3.1.0
info:
  title: Contratto API dell'assistente RAG per il troubleshooting
  version: 0.4.0
  description: >
    Contratto API per un assistente Retrieval-Augmented Generation (RAG) dedicato al troubleshooting di applicazioni backend a microservizi.
    Il sistema riceve una domanda tecnica, un eventuale contesto dell'incidente ed eventuali evidenze di telemetria normalizzate. Recupera quindi informazioni pertinenti dalla Knowledge Base e genera una risposta corredata dalle fonti tecniche utilizzate.
    
paths:
  /troubleshoot:
    post:
      summary: Analizza un incidente tecnico tramite il sistema RAG
      description: >
        Riceve una domanda relativa a un problema tecnico, interroga la Knowledge Base attraverso la pipeline di retrieval configurata e genera una risposta basandosi sulle fonti documentali recuperate.
      operationId: troubleshootIncident
      requestBody:
        required: true
        description: |
          Informazioni necessarie per effettuare l'analisi di troubleshooting.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TroubleshootingRequest'
            examples:
              simpleQuestion:
                summary: Domanda senza telemetria strutturata
                value:
                  question: >
                    Il servizio checkout riceve connection refused durante il pagamento, quale potrebbe essere il problema?
                  incident_context: >
                    Durante il checkout la chiamata verso il servizio payment fallisce.
                  service: checkout
              telemetryQuestion:
                summary: Domanda con evidenze di telemetria
                value:
                  question: |
                    Perché il checkout non riesce a completare il pagamento?
                  incident_context: |
                    Il problema si verifica durante la fase di charge.
                  service: checkout
                  telemetry:
                    trace_id: trace-example-001
                    logs:
                      - service: checkout
                        severity: ERROR
                        message: connection refused while contacting payment
                    spans:
                      - service: checkout
                        operation: payment/charge
                        status: ERROR
                        peer_service: payment
                    metrics: []
      responses:
        '200':
          description: Analisi completata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TroubleshootingResponse'
              examples:
                successfulResponse:
                  value:
                    answer: >
                      Le evidenze indicano che il servizio checkout non riesce a raggiungere il servizio payment.
                       Verificare la disponibilità del servizio e la configurazione dell'endpoint utilizzato.
                    sources:
                      - citation_id: FONTE_1
                        document_id: document-payment-unreachable
                        source: payment-unreachable.md
                        chunk_id: payment-unreachable-0001
                        document_type: runbook
                        section: Possible causes
                        service: payment
                        category: service-unavailable
                        rank: 1
                        retriever: rrf
                        score: null
                        fused_score: 0.0328
                        reranker_score: 4.82
                    latency_ms: 842
                    operational_metrics:
                      retrieval_latency_ms: 74
                      reranking_latency_ms: 31
                      prompt_build_latency_ms: 2
                      generation_latency_ms: 721
                      total_latency_ms: 842
                      token_usage:
                        input_tokens: 615
                        output_tokens: 126
                        total_tokens: 741
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  schemas:
    TroubleshootingRequest:
      type: object
      description: |
        Informazioni fornite dall'utente per analizzare un incidente tecnico.
      additionalProperties: false
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
          description: |
            Domanda tecnica alla quale l'assistente deve rispondere.
          examples:
            - >-
              La chiamata al servizio payment restituisce HTTP 500. Quale può essere la causa?
        incident_context:
          type:
            - string
            - 'null'
          default: null
          description: >
            Descrizione testuale opzionale del contesto dell'incidente.
          examples:
            - >-
              Il servizio checkout riceve connection refused dal servizio payment.
        service:
          type:
            - string
            - 'null'
          default: null
          description: >
            Nome del microservizio principalmente interessato dall'incidente, se noto.
          examples:
            - checkout
        telemetry:
          anyOf:
            - $ref: '#/components/schemas/TelemetryContext'
            - type: 'null'
          default: null
          description: >
            Evidenze di telemetria normalizzate relative all'incidente.
    TelemetryContext:
      type: object
      additionalProperties: false
      description: |
        Contesto di telemetria normalizzato associato all'incidente analizzato.
      properties:
        trace_id:
          type:
            - string
            - 'null'
          default: null
          description: >
            Identificativo della distributed trace associata all'incidente, se disponibile.
        logs:
          type: array
          default: []
          description: |
            Evidenze provenienti dai log applicativi.
          items:
            $ref: '#/components/schemas/LogEvidence'
        spans:
          type: array
          default: []
          description: |
            Span rilevanti estratti dalla distributed trace.
          items:
            $ref: '#/components/schemas/SpanEvidence'
        metrics:
          type: array
          default: []
          description: |
            Metriche considerate rilevanti per l'incidente.
          items:
            $ref: '#/components/schemas/MetricEvidence'
    LogEvidence:
      type: object
      additionalProperties: false
      required:
        - service
        - message
      description: |
        Evidenza normalizzata proveniente da un log applicativo.
      properties:
        timestamp:
          type:
            - string
            - 'null'
          format: date-time
          default: null
          description: 'Timestamp del log, se disponibile.'
        service:
          type: string
          minLength: 1
          description: |
            Microservizio che ha prodotto il log.
        severity:
          type:
            - string
            - 'null'
          default: null
          description: |
            Livello di severità del log.
          examples:
            - ERROR
        message:
          type: string
          minLength: 1
          description: |
            Messaggio di log normalizzato.
          examples:
            - connection refused while contacting payment
        error_type:
          type:
            - string
            - 'null'
          default: null
          description: |
            Tipo di errore o eccezione, se disponibile.
          examples:
            - ConnectionRefusedError
    SpanEvidence:
      type: object
      additionalProperties: false
      required:
        - service
        - operation
        - status
      description: |
        Evidenza proveniente da uno span di una distributed trace.
      properties:
        span_id:
          type:
            - string
            - 'null'
          default: null
          description: |
            Identificativo dello span, se disponibile.
        service:
          type: string
          minLength: 1
          description: |
            Microservizio associato allo span.
        operation:
          type: string
          minLength: 1
          description: |
            Operazione rappresentata dallo span.
          examples:
            - payment/charge
        status:
          type: string
          enum:
            - OK
            - ERROR
            - UNSET
          description: |
            Stato normalizzato dello span.
        peer_service:
          type:
            - string
            - 'null'
          default: null
          description: |
            Eventuale servizio remoto coinvolto nell'operazione.
          examples:
            - payment
        error_message:
          type:
            - string
            - 'null'
          default: null
          description: |
            Messaggio di errore associato allo span, se disponibile.
    MetricEvidence:
      type: object
      additionalProperties: false
      required:
        - name
        - value
      description: |
        Evidenza normalizzata derivata da una metrica.
      properties:
        service:
          type:
            - string
            - 'null'
          default: null
          description: |
            Microservizio al quale la metrica è associata.
        name:
          type: string
          minLength: 1
          description: |
            Nome della metrica.
          examples:
            - request.duration
        value:
          type: number
          description: |
            Valore osservato della metrica.
        unit:
          type:
            - string
            - 'null'
          default: null
          description: |
            Unità di misura della metrica.
          examples:
            - ms
    TroubleshootingResponse:
      type: object
      description: |
        Risultato dell'analisi di troubleshooting.
      additionalProperties: false
      required:
        - answer
        - sources
        - latency_ms
        - operational_metrics
      properties:
        answer:
          type: string
          minLength: 1
          description: >
            Risposta generata dal sistema sulla base della domanda, delle
            evidenze disponibili e dei documenti recuperati.
        sources:
          type: array
          description: >
            Fonti documentali utilizzate per costruire la risposta.

            L'array può essere vuoto, ad esempio nel caso del baseline
            sperimentale LLM senza RAG.
          items:
            $ref: '#/components/schemas/SourceReference'
        latency_ms:
          type: number
          format: double
          minimum: 0
          description: >
            Tempo totale di elaborazione della richiesta espresso in
            millisecondi. Il valore coincide con total_latency_ms.
        operational_metrics:
          $ref: '#/components/schemas/OperationalMetrics'
          description: |
            Dettaglio delle latenze dei singoli stadi e dell'utilizzo dei token.
    OperationalMetrics:
      type: object
      additionalProperties: false
      required:
        - retrieval_latency_ms
        - reranking_latency_ms
        - prompt_build_latency_ms
        - generation_latency_ms
        - total_latency_ms
      description: |
        Metriche operative raccolte durante l'elaborazione della richiesta.
      properties:
        retrieval_latency_ms:
          type: number
          format: double
          minimum: 0
          description: |
            Tempo impiegato dal recupero documentale, espresso in millisecondi.
        reranking_latency_ms:
          type: number
          format: double
          minimum: 0
          description: |
            Tempo impiegato dal secondo stadio di reranking, espresso in millisecondi.
        prompt_build_latency_ms:
          type: number
          format: double
          minimum: 0
          description: |
            Tempo impiegato per costruire il prompt, espresso in millisecondi.
        generation_latency_ms:
          type: number
          format: double
          minimum: 0
          description: |
            Tempo impiegato dal modello generativo, espresso in millisecondi.
        total_latency_ms:
          type: number
          format: double
          minimum: 0
          description: |
            Tempo complessivo di elaborazione, espresso in millisecondi.
        token_usage:
          anyOf:
            - $ref: '#/components/schemas/TokenUsage'
            - type: 'null'
          default: null
          description: |
            Conteggio dei token comunicato dal provider LLM, se disponibile.
    TokenUsage:
      type: object
      additionalProperties: false
      required:
        - input_tokens
        - output_tokens
        - total_tokens
      description: |
        Numero di token elaborati e generati dal modello LLM.
      properties:
        input_tokens:
          type: integer
          minimum: 0
          description: |
            Token elaborati nel prompt.
        output_tokens:
          type: integer
          minimum: 0
          description: |
            Token prodotti nella risposta.
        total_tokens:
          type: integer
          minimum: 0
          description: |
            Somma dei token di input e di output.
    SourceReference:
      type: object
      additionalProperties: false
      required:
        - citation_id
        - document_id
        - source
        - chunk_id
        - document_type
        - rank
        - retriever
      description: >
        Riferimento alla porzione della Knowledge Base utilizzata durante la
        generazione della risposta.
      properties:
        citation_id:
          type: string
          pattern: '^FONTE_[1-9][0-9]*$'
          description: |
            Identificativo fornito al modello e restituito al client per citare questa fonte.
          examples:
            - FONTE_1
        document_id:
          type: string
          minLength: 1
          description: |
            Identificativo stabile del documento nella Knowledge Base.
          examples:
            - document-payment-unreachable
        source:
          type: string
          minLength: 1
          description: |
            Documento originale dal quale proviene il contenuto recuperato.
          examples:
            - payment-unreachable.md
        chunk_id:
          type: string
          minLength: 1
          description: |
            Identificativo stabile del chunk recuperato.
          examples:
            - payment-unreachable-0001
        document_type:
          type: string
          minLength: 1
          description: |
            Tipologia del documento dal quale proviene il chunk.
          examples:
            - runbook
        section:
          type:
            - string
            - 'null'
          default: null
          description: |
            Sezione del documento dalla quale proviene il chunk.
          examples:
            - Possible causes
        service:
          type:
            - string
            - 'null'
          default: null
          description: |
            Microservizio associato alla fonte, se disponibile.
          examples:
            - payment
        category:
          type:
            - string
            - 'null'
          default: null
          description: |
            Categoria documentale associata alla fonte, se disponibile.
          examples:
            - service-unavailable
        rank:
          type: integer
          minimum: 1
          description: |
            Posizione della fonte nella graduatoria finale utilizzata dal prompt.
        retriever:
          type: string
          minLength: 1
          description: |
            Componente che ha prodotto la graduatoria finale.
          examples:
            - rrf
        score:
          type:
            - number
            - 'null'
          default: null
          description: |
            Punteggio originale assegnato dal retriever, se disponibile.
        fused_score:
          type:
            - number
            - 'null'
          default: null
          description: |
            Punteggio assegnato dalla fusione delle graduatorie, se disponibile.
        reranker_score:
          type:
            - number
            - 'null'
          default: null
          description: |
            Punteggio assegnato dal secondo stadio di reranking, se disponibile.
    ErrorResponse:
      type: object
      description: |
        Errore restituito quando la richiesta non può essere completata.
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          minLength: 1
          description: >
            Codice applicativo stabile e leggibile automaticamente che
            identifica l'errore.
          examples:
            - LLM_SERVICE_UNAVAILABLE
        message:
          type: string
          minLength: 1
          description: |
            Descrizione dell'errore in forma leggibile.
          examples:
            - Il servizio LLM remoto non è raggiungibile.
  responses:
    BadRequest:
      description: |
        La richiesta è semanticamente non valida o non può essere elaborata.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_REQUEST
            message: La richiesta non può essere elaborata.
    ValidationError:
      description: |
        Il corpo della richiesta non rispetta il contratto API previsto.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Il campo question è obbligatorio.
    InternalServerError:
      description: >
        Si è verificato un errore interno durante l'elaborazione della
        richiesta.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Errore interno durante l'elaborazione della richiesta.
    BadGateway:
      description: |
        Il servizio LLM remoto ha restituito una risposta non valida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: LLM_INVALID_RESPONSE
            message: Il servizio LLM remoto ha restituito una risposta non valida.
    ServiceUnavailable:
      description: >
        Una dipendenza necessaria al completamento della richiesta non è
        disponibile.

        Può verificarsi, ad esempio, quando il servizio LLM remoto non è
        raggiungibile.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: LLM_SERVICE_UNAVAILABLE
            message: Il servizio LLM remoto non è raggiungibile.